Try to have a faster implementation for set
